+++ src/Symfony/Bundle/FrameworkBundle/Test/WebTestCase.php
@@ -23,6 +23,7 @@ abstract class WebTestCase extends KernelTestCase
 {
     use ForwardCompatTestTrait;
     use WebTestAssertionsTrait;
+    use MailerAssertionsTrait;
 
     private function doTearDown()
     {